Container user is now the same as the host user
It is now possible to specify the user ID and group ID of the host
system user as environment variables when starting the container. During
startup, the container will create a user with the same user ID and
group ID, and run ComfyUI with this user. This is helpful, because the
model and custom node files downloaded by the ComfyUI Manager will be
owned by the host system user instead of root. The documentation in the
read me was updated to reflect this change.

Furthermore, while the custom nodes themselves are installed outside of
the container, their requirements are installed inside of the container.
This means that stopping and removing the container will remove the
installed requirements. Therefore, the dependencies are now installed
when the container is started again.