Added Missing Dependencies
Installed the packages libgl1-mesa-glx and libglib2.0-0, which will
hopefully fix issues with the ComfyUI Impact Pack, ComfyUI Impact
Subpack, and ComfyUI Inspire Pack custom nodes that were missing the
libGL.so.1 and libgthread-2.0.so.0 libraries.

Also, the APT cache is now being cleaned after installing the packages
to reduce the image size and the changelog was updated to reflect the
changes.