Updated All Dependencies
Updated the base image of the Dockerfile to the latest version of
PyTorch (from PyTorch 2.7.1, CUDA 12.6, and cuDNN 9 to PyTorch 2.8.0,
CUDA 12.9, and cuDNN 9).

Updated the ComfyUI version to the latest version (from ComfyUI 0.3.40
to ComfyUI 0.3.49).

Updated the ComfyUI Manager version to the latest version (from ComfyUI
Manager 3.33 to ComfyUI Manager 3.35).